+++ java-service/src/main/java/com/ethsimulator/market/YieldQuote.java
@@ -2,11 +2,12 @@ package com.ethsimulator.market;
 
 import com.ethsimulator.simulation.RiskTier;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record YieldQuote(
         String protocol,
-        double apyPct,
+        BigDecimal apyPct,
         String source,
         RiskTier riskTier,
         Instant observedAt
